Get tests to pass

+++ dpc-attribution/src/test/java/gov/cms/dpc/attribution/jdbi/RelationshipDAOUnitTest.java
@@ -11,7 +11,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.lang.reflect.Field;
-import java.math.BigInteger;
 import java.util.Arrays;
 import java.util.List;
 
@@ -135,7 +134,7 @@ class RelationshipDAOUnitTest extends AbstractAttributionDAOTest {
 
 		Session session = db.getSessionFactory().getCurrentSession();
 		String sql = "select increment_by from pg_sequences where sequencename = 'attributions_id_seq'";
-		int dbIncrement = ((BigInteger) session.createNativeQuery(sql).getSingleResult()).intValue();
+		int dbIncrement = ((Long) session.createNativeQuery(sql).getSingleResult()).intValue();
 
 		assertEquals(hibernateIncrement, dbIncrement);
 	}
